server_python/pyproject.toml
[project]
name = "coptic-translator-backend"
version = "0.1.0"
description = "Backend server for Coptic translation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "slowapi>=0.1.9",
    "googletrans-py>=4.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "anthropic>=0.71.0",
    "openai>=2.6.1",
]

[project.scripts]
server = "main:app"
